Update gem5Run to use factory pattern
Instead of using inheiritance to have two different constructors for
gem5Run (one for FS and one for SE), use the factory pattern. This
greatly simplifies the code to load from the database fixing a bug where
some members of the class were overwritten by the constructor when
loading from the database.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>