Fix celery users (#50)
* Remove extra files from run directory

* Allow run to execute in other directory

This change updates the outdir of run to allow the user to specify a
full path and updates run() with another parameter, cwd.

* Update the task to use the new cwd parameter

This change updates the run_gem5_instance task to take an optional cwd
parameter. This allows the user to specify which directory to run the
gem5 task in.