Generalize the ArtifactDB object
This puts on the path to being able to implement other database types
(e.g., a mock database as mentioned in DARCHR-115).

This is working towards DARCHR-121.

This patch makes ArtifactDB fully virtual and removes it from the global
namespace exported by Artifact.

`getDBConnection` now is correctly implemented as a singleton factory.

It also adds a searchByName and searchByType functions to the abstract
DB and implementes them in the MongoDB version to move the
mongo-specific code from artifact.

The documentation is also updated to always use getDBConnection.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>