Uploading artifact before registering artifact
Currently, we register an artifact with the database before it is being uploaded. This leads to the case where the artifact is failed to be uploaded, and the file won't be uploaded in later runs, since it is registered to the database. This would cause an issue if we want to retrieve the file later (e.g. querying `result` artifacts).

This commit fixes that by uploading the file first, and registers the corresponding artifact later.