run: don't fail if object missing from db
In loadFromDict, if there is an artifact that's not found in the
database, the current behavior is to raise an exception. This change
instead inserts a "placeholder" dictionary with the original UUID and a
note that the data can't be found. Currently, this function is only used
when extracting runs from the database.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>