Add ability to configure db connection
This change updates the db constructor to take a parameter to set up the
db connection. It also changes the interface for getDBConnection to take
a "URI" for the database connection. Right now, the only option is
MongoDB, but in the future, we can add flat files or other database
backends easily.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>